Detect the hash in use before downloading so it can be put into the URL.
This moves the logic to detect the type of hash function before the
download, as that way the name of the hash function can be part of the
download URL.

In order to avoid reproducing the table of hash functions twice,
we abuse the fact that ${hashtype} is always the first part of the
summing program (e.g. md5 has md5sum), and invoke ${hashtype}sum.

It also switches to a case statement instead of a long chain of conditionals.
The default case currently simply throws an error (because the hash function
wasn't recognized).