Add support for automatically extracting the namespace from .git/config
This should make it possible to support more namespaces without having
to hardcode them in rfpkg-minimal. I worry my parsing is a little
fragile (if the format of .git/config were to change, it would need
fixing)-- but that's probably better than having to add new namespaces
by hand and need to change things anyway.